Remove set_value_optimized_out.
This patch undoes a recent change by Pedro:
  http://sourceware.org/ml/gdb-patches/2013-07/msg00178.html

however, I hope that later patches in this series will move us back to a
situation with which Pedro will be happy.  For now, I believe the current
code is functional, and was only rejected by Pedro on the grounds that /if/
in the future we added support for values being both lazy /and/ marked
optimized out then this would break.  As I said, future patches will move
us back in that direction.

Other that that not really a contentious patch, just messing with the API
in order to make available the information about which parts of the value
we'd like optimized out.

gdb/ChangeLog

2013-08-08  Andrew Burgess  <aburgess@broadcom.com>

	* ada-lang.c (coerce_unspec_val_to_type): Switch to
	mark_value_bytes_optimized_out, and only mark non-lazy values as
	optimized out.
	* dwarf2loc.c (read_pieced_value): Switch to use
	mark_value_bytes_optimized_out, or mark_value_bits_optimized_out.
	(write_pieced_value): Switch to use
	mark_value_bytes_optimized_out.
	* findvar.c (read_frame_register_value): Switch to use
	mark_value_bytes_optimized_out.
	(value_from_register): Likewise.
	* value.c (allocate_optimized_out_value, value_fetch_lazy): Switch
	to use mark_value_bytes_optimized_out.
	(value_optimized_out_const, set_value_optimized_out): Remove.
	(mark_value_bytes_optimized_out, mark_value_bits_optimized_out):
	New function definition.
	* value.h (value_optimized_out_const, set_value_optimized_out):
	Remove.
	(mark_value_bytes_optimized_out, mark_value_bits_optimized_out):
	Add new declaration.