Mark optimized out value as non-lazy.
This is a re-posting of this patch:
  http://sourceware.org/ml/gdb-patches/2013-07/msg00058.html

I don't think that patch was ever rejected, it more just ran out of steam.
Pedro noticed that we could go beyond this patch and do more, even
releasing the value contents when we spot values are fully optimized out.
I agree, but believe that is a separate idea, that can always be added
later, for this case we've not even allocated any value contents yet.

gdb/ChangeLog

	* value.c (allocate_optimized_out_value): Mark fully optimized out
	values as non-lazy.