Display <optimized out> for optimized out register values.
This makes the display of register values consistent.  Currently, printing
a register that is not saved (maybe due to the ABI), will print "<optimized
out>" when printed as "p/x $reg", but will print "*value not available*"
when printed with "info register reg".   After this patch "<optimized out>"
is used throughout.

gdb/ChangeLog

	* infcmd.c (default_print_one_register_info): Add detection of
	optimized out values.
	(default_print_registers_info): Switch to using
	get_frame_register_value.

gdb/testsuite/ChangeLog

	* gdb.dwarf2/dw2-reg-undefined.exp: Change pattern for info
	register to "<optimized out>", and also print the registers.